moved the expenses and providers state via get request up a level to the mainContainer so we can rewnder the ExpensePieChart on the homepage.
